refactor: move list-box core and base CSS to styles folder
