+++ packages/avatar-group/src/styles/vaadin-avatar-group-menu-core-styles.js
@@ -3,6 +3,6 @@
  * Copyright (c) 2020 - 2025 Vaadin Ltd.
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
-import { listBoxStyles } from '@vaadin/list-box/src/vaadin-list-box-core-styles.js';
+import { listBoxStyles } from '@vaadin/list-box/src/styles/vaadin-list-box-core-styles.js';
 
 export { listBoxStyles as avatarGroupMenuStyles };
