+++ packages/list-box/src/styles/vaadin-list-box-base-styles.d.ts
