+++ packages/list-box/src/styles/vaadin-list-box-core-styles.js
